Discord Elixir v1.0.0 DiscordElixir.RealtimeClient.Helpers

Writing a Bot? These helpers should help make it easier.

Functions

actionable_message_for?(username, payload, state)

Specs

actionable_message_for?(String.t, map, map) :: boolean

This checks that an incoming message is private or is a mention to the defined user

current_channels(state, type \\ "private")

Specs

current_channels(map, String.t) :: list

Get the current channels of a specific type for a client.

msg_command_parse(payload)

Specs

msg_command_parse(map) :: {String.t, String.t}

Parses a message command from payload which is content leading with ‘!’