Discord Elixir v1.0.0 DiscordElixir.RestClient.Resources.Channel

Convience helper for channel

Functions

delete(conn, channel_id)

Specs

delete(pid, number) :: map

Delete a guild channel, or close a private message

Requires the ‘MANAGE_GUILD’ permission for the guild containing the channel.

Parameters

conn: User connection for REST holding auth info

channel_id: Channel id

Examples

Channel.delete(conn, 93209293090239)

get(conn, channel_id)

Specs

get(pid, number) :: map

Get channel

Parameters

conn: User connection for REST holding auth info

channel_id: Channel id

Examples

Channel.get(conn, 99999999993832)

messages(conn, channel_id)

Specs

messages(pid, number) :: map

Retrieve messages for a channel

Requires the ‘READ_MESSAGES’ permission for the guild containing the channel.

Parameters

conn: User connection for REST holding auth info

channel_id: Channel id

Examples

Channel.messages(conn, 439093409304934)

modify(conn, channel_id, options)

Specs

modify(pid, number, map) :: map

Update a channels settings

Requires the ‘MANAGE_GUILD’ permission for the guild containing the channel.

Parameters

conn: User connection for REST holding auth info

channel_id: Channel id

options: Updateable options which include (name, position, topic, bitrate)

Examples

Channel.modify(conn, 3290293092093, %{name: "my-channel", topic" "we all are friends here"})

send_file(conn, channel_id, file_data)

Specs

send_file(pid, number, map) :: map

Post a message to a guild text or DM channel

Requires the ‘SEND_MESSAGES’ permission to be present on the current user.

Parameters

conn: User connection for REST holding auth info

channel_id: Channel id

file_data: filed data which include (content, nonce, tts, file)

Examples

Channel.send_messages(conn, 439093409304934, %{content: "Check this out!", file: "/local/path/to/file.jpg"})

send_message(conn, channel_id, message_data)

Specs

send_message(pid, number, String.t) :: map

Post a message to a guild text or DM channel

Requires the ‘SEND_MESSAGES’ permission to be present on the current user.

Parameters

conn: User connection for REST holding auth info

channel_id: Channel id

message_data: message data which include (content, nonce, tts)

Examples

Channel.send_message(conn, 439093409304934, %{content: "Hi! Friends!"})