Discord Elixir v1.1.18 DiscordEx.Client.Utility

Utilty methods to be used for discord clients.

Normalizers, Encoders, and Decoders

Functions

agent_update(agent, n)

agent_update(pid(), any()) :: nil

Generic function for updating the value of an agent process

agent_value(agent)

agent_value(pid()) :: any()

Generic function for getting the value from an agent process

normalize_atom(atom)

normalize_atom(atom()) :: String.t()

Convert atom to string

opcode(codes, value)

opcode(map(), integer()) :: atom()

opcode(map(), atom()) :: integer()

Get the atom value of and opcode using an integer value

payload_build(op, data, seq_num \\ nil, event_name \\ nil)

payload_build(number(), map(), number(), String.t()) :: binary()

Build a binary payload for discord communication

payload_build_json(op, data, seq_num \\ nil, event_name \\ nil)

payload_build_json(number(), map(), number(), String.t()) :: binary()

Build a json payload for discord communication

payload_decode(codes, arg)

payload_decode(list(), {atom(), binary()}) :: map()

payload_decode(list(), {atom(), binary()}) :: map()

Decode json payload received from discord into a map