Discord Elixir v1.1.18 DiscordEx.Connections.REST

Discord uses a REST interface to send data to the API.

Types

body()

body() :: binary() | {:form, [{atom(), any()}]} | {:file, binary()}

headers()

headers() ::
  [{binary(), binary()}] |
  %{optional(binary()) => binary()}

Functions

delete(url, headers \\ [], options \\ [])

delete(binary(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()} |
  {:error, HTTPoison.Error.t()}

Issues a DELETE request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

delete!(url, headers \\ [], options \\ [])

delete!(binary(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() |
  HTTPoison.AsyncResponse.t()

Issues a DELETE request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

get(url, headers \\ [], options \\ [])

get(binary(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()} |
  {:error, HTTPoison.Error.t()}

Issues a GET request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

get!(url, headers \\ [], options \\ [])

get!(binary(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() |
  HTTPoison.AsyncResponse.t()

Issues a GET request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

head(url, headers \\ [], options \\ [])

head(binary(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()} |
  {:error, HTTPoison.Error.t()}

Issues a HEAD request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

head!(url, headers \\ [], options \\ [])

head!(binary(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() |
  HTTPoison.AsyncResponse.t()

Issues a HEAD request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

options(url, headers \\ [], options \\ [])

options(binary(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()} |
  {:error, HTTPoison.Error.t()}

Issues an OPTIONS request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

options!(url, headers \\ [], options \\ [])

options!(binary(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() |
  HTTPoison.AsyncResponse.t()

Issues a OPTIONS request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

patch(url, body, headers \\ [], options \\ [])

patch(binary(), body(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()} |
  {:error, HTTPoison.Error.t()}

Issues a PATCH request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

patch!(url, body, headers \\ [], options \\ [])

patch!(binary(), body(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() |
  HTTPoison.AsyncResponse.t()

Issues a PATCH request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

post(url, body, headers \\ [], options \\ [])

post(binary(), body(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()} |
  {:error, HTTPoison.Error.t()}

Issues a POST request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

post!(url, body, headers \\ [], options \\ [])

post!(binary(), body(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() |
  HTTPoison.AsyncResponse.t()

Issues a POST request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

process_url(target_url)

put(url, body, headers \\ [], options \\ [])

put(binary(), body(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()} |
  {:error, HTTPoison.Error.t()}

Issues a PUT request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

put!(url, body, headers \\ [], options \\ [])

put!(binary(), body(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() |
  HTTPoison.AsyncResponse.t()

Issues a PUT request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

request(method, url, body \\ "", headers \\ [], options \\ [])

request(atom(), binary(), body(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()} |
  {:error, HTTPoison.Error.t()}

Issues an HTTP request with the given method to the given url.

This function is usually used indirectly by get/3, post/4, put/4, etc

Args:

method - HTTP method as an atom (:get, :head, :post, :put, :delete, etc.)

url - target url as a binary string or char list

body - request body. See more below

headers - HTTP headers as an orddict (e.g., [{"Accept", "application/json"}])

options - Keyword list of options

Body:

binary, char list or an iolist

{:form, [{K, V}, ...]} - send a form url encoded

{:file, "/path/to/file"} - send a file

Options:

:timeout - timeout to establish a connection, in milliseconds. Default is 8000

:recv_timeout - timeout used when receiving a connection. Default is 5000

:stream_to - a PID to stream the response to

:proxy - a proxy to be used for the request; it can be a regular url or a {Host, Proxy} tuple

:proxy_auth - proxy authentication {User, Password} tuple

:ssl - SSL options supported by the ssl erlang module

:follow_redirect - a boolean that causes redirects to be followed

:max_redirect - an integer denoting the maximum number of redirects to follow

:params - an enumerable consisting of two-item tuples that will be appended to the url as query string parameters

Timeouts can be an integer or :infinity

This function returns {:ok, response} or {:ok, async_response} if the request is successful, {:error, reason} otherwise.

Examples

request(:post, "https://my.website.com", "{\"foo\": 3}", [{"Accept", "application/json"}])

request!(method, url, body \\ "", headers \\ [], options \\ [])

request!(atom(), binary(), body(), headers(), Keyword.t()) :: HTTPoison.Response.t()

Issues an HTTP request with the given method to the given url, raising an exception in case of failure.

request!/5 works exactly like request/5 but it returns just the response in case of a successful request, raising an exception in case the request fails.

start()

Starts HTTPoison and its dependencies.