Discord Elixir v1.1.18 DiscordEx.Voice.Client

This client is for specifically working with voice. You can pass this process to your regular client if you wish to use it with your bot.

Examples

token = "<your-token>"
DiscordEx.Voice.Client.connect(base_client, %{guild_id: 392090239, channel_id: 23208203092390)
#=> {:ok, #PID<0.180.0>}

Functions

connect(base_client, options \\ %{})

connect(pid(), map()) :: {:ok, pid()}

Initialize a voice connection

disconnect(voice_client)

disconnect(pid()) :: atom()

Kill a voice connection

handle_event(arg, state)

identify(state)

init(state)

Callback implementation for c::websocket_client.init/1.

onconnect(WSReq, state)

Callback implementation for c::websocket_client.onconnect/2.

ondisconnect(arg, state)

Callback implementation for c::websocket_client.ondisconnect/2.

opcodes()

reconnect(client_pid, options)

reconnect(pid(), map()) :: {:ok, pid()}

Reconnect or initiate voice connection

socket_url(url)

socket_url(map()) :: String.t()

start_link(opts)

websocket_handle(arg, socket, state)

Callback implementation for c::websocket_client.websocket_handle/3.

websocket_info(arg1, arg2, state)

Ability to update speaking state

websocket_terminate(reason, conn_state, state)

Callback implementation for c::websocket_client.websocket_terminate/3.