Discord Elixir v1.1.18 DiscordEx.RestClient.Resources.Image

Convience helper for images

Functions

avatar_url(user_id, avatar_id)

avatar_url(String.t(), String.t()) :: String.t()

Get the user avatar URL

Parameters

user_id: The user_id to retrieve the avatar URL for

avatar_id: The avatar_id for the user to retrieve the avatar URL for

Examples

Image.avatar_url("99999999993832","f3e8329c329020329")

icon_url(server_id, icon_id)

icon_url(String.t(), String.t()) :: String.t()

Get the guild icon URL

Parameters

server_id: The server_id to retrieve the icon URL for

icon_id: The icon_id for the user to retrieve the icon URL for

Examples

Image.icon_url("99999999993832","f3e8329c329020329")